refactor media factory
